[SDAN-631] The 'PRINT REPORT' action produces a page without any report data (#945)
