+++ assets/company-reports/components/index.js
@@ -1,10 +1,13 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { panels } from '../utils';
+import {get} from 'lodash';
 
+import {createStore, render} from 'utils';
+import { panels } from '../utils';
+import companyReportReducer from '../reducers';
+
+const store = createStore(companyReportReducer, 'CompanyReports');
 const Panel = panels[window.report];
 
-ReactDOM.render(
-    <Panel data={window.reportData} print={true}/>, document.getElementById('print-reports')
-);
-
+render(store, Panel, document.getElementById('print-reports'), {
+    results: get(window, 'reportData.results'),
+    print: true
+});
