+++ assets/company-reports/utils.js
@@ -4,16 +4,7 @@ import CompanySavedSearches from './components/CompanySavedSearches';
 import ProductStories from './components/ProductStories';
 import Company from './components/Company';
 import SubscriberActivity from './components/SubscriberActivity';
-
-
-export const REPORTS_NAMES = {
-    'COMPANY_SAVED_SEARCHES': 'company-saved-searches',
-    'USER_SAVED_SEARCHES': 'user-saved-searches',
-    'COMPANY_PRODUCTS': 'company-products',
-    'PRODUCT_STORIES': 'product-stories',
-    'COMPANY': 'company',
-    'SUBSCRIBER_ACTIVITY': 'subscriber-activity'
-};
+import {REPORTS_NAMES} from './actions';
 
 export const panels = {
     [REPORTS_NAMES.COMPANY_SAVED_SEARCHES]: CompanySavedSearches,
