+++ newsroom/templates/reports_print.html
@@ -10,6 +10,7 @@
 {% block script %}
 
     <script>
+        window.newsroom = {{ get_client_config() | tojson | safe }};
         reportData = {{ data | tojson | safe }};
         report = {{ report | tojson | safe }};
     </script>
